requirements-rpi.txt
Flask
python-dotenv
Adafruit_DHT
RPi.GPIO
